Style(minjae/develop): map 컴포넌트 함수, 스타일 분류
