+++ src/style/map.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const MapWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+`;
